+++ README.md
@@ -13,4 +13,5 @@ then do this two step in two seperated terminal windows:
 npm run react-dev
 npm run server-dev
 
-//
+we must do the react router
+npm install react-router --save
