+++ package.json
@@ -15,8 +15,7 @@
     "babel-loader": "^6.3.2",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
-    "webpack": "^2.2.1",
-    "webpack-cli": "^2.0.15"
+    "webpack": "^2.2.1"
   },
   "dependencies": {
     "angular": "^1.6.3",
@@ -27,6 +26,7 @@
     "mongoose": "^4.8.6",
     "mysql": "^2.13.0",
     "react": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "react-dom": "^15.4.2",
+    "react-router": "^4.2.0"
   }
 }
