+++ react-client/src/Jozaa/components/Password.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
-//react
 //the style for the password
 const passwordStyle={
-  //marginTop:'100px',
   color:'red',
   fontWeight:'bold',
   textAlign:'center'
@@ -10,15 +8,17 @@ const passwordStyle={
 //the style for the password input
 const userInputStyle={
   padding:'3px',
+  //this three to make it center
   display: 'block',
   marginRight: 'auto',
   marginLeft: 'auto',
-};      
+};  
+//what render for password component
 const Password = (props) => (
   <div>
     <h3 style={passwordStyle}> Password: </h3>
     <input placeholder="Insert your password" style={userInputStyle} ></input>
   </div>
 )
-
+//export this component to can use
 export default Password;
